Add files via upload
Changes variable names GPIO1, GPIO2 and GPIO3 to gpio1, gpio2 and gpio3 to mitigate problem compiling for an Arduino Nano Every in IDE 2.3.2. 
---------------------------------------------------------------------------------
If you look in the header files included with avr-gcc/7.3.0-atmel3.6.1-arduino7, you will find that GPIO1, GPIO2 and GPIO3 are used therein.

Simply rename GPIO1, GPIO2 and GPIO3 to gpio1, gpio2 and gpio3 in the RDA5807 library and it will compile.